media: dvb: Fix handling of multiple PIDs per recording filter
When starting feed fails while the feed is part of a filter that
holds multiple PIDs for recording, the filter's state does not
change and therefore the feeds which were started were not stopped
and de-allocated properly. In addition the dvr-feed count
is per feed basis and should be decremented for each feed that
is stopped part of the recording filter.

Change-Id: I739e6d40596df3c80cd0843f89eb4a8d3d51ddf4
Signed-off-by: Hamad Kadmany <hkadmany@codeaurora.org>