USB: android: Wakeup both reader and writer processes upon TX/RX error
The writer process is blocked on a wait queue if there are no free
TX requests.  It is unblocked from TX completion handler. When error
is encountered on RX path, wakeup the writer thread.  If the host
stops reading the data after RX errors, the writer process will never
be unblocked.  Similarly wakeup reader process when error is encountered
on TX path.

CRs-Fixed: 449129
Change-Id: I4fa133d293c96b629d1506a0054b048e9551abd1
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>