i2c_qup: always set rate and log a message when fail
Clock set rate is failing silently when clock rate is not
available. Additionally clock rate is not set when no rate
is specified in platform data. This patch always set the
clock's rate and logs a message if it fails to set rate
or if no rate is set by platfrom data.

CRs-Fixed: 452047
Change-Id: I0551a12ca2861f8e1a48ee344d6ffcf898c96971
Signed-off-by: Gilad Avidov <gavidov@codeaurora.org>