msm: vidc: Change order of operations
The video driver needs to call into IOMMU driver which then calls
into secure environment to unlock access to IOMMU bfb registers before
the video driver brings the video core out of reset. This is
because the IOMMU driver needs to program the IOMMU bfb registers
before the video core is brought out of reset.

Signed-off-by: Vinay Kalia <vkalia@codeaurora.org>
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>
Change-Id: Ica80eaec53e8512b7d614f31eca33ca2908f3722