mmc: msm_sdcc: always check sdcc state first
Currently bam is reset without checking the current state of sdcc clocks.
If clocks are not on when bam reset is attempted then this would lead to
a system crash due to unclocked register access.

In order to fix this, always check the state of the sdcc hardware before
starting any request. This way even if a new request comes in without the
clocks being on, it would not go through.

CRs-Fixed: 453099
Change-Id: I872aa60d615617bd448b977cdf06ffcf900e111e
Signed-off-by: Krishna Konda <kkonda@codeaurora.org>