arm/dt: msm8226: Fix PM8226 S1 voltage setting to nominal
The current setting of 1.2875 V is not a supported
PMIC FTS setpoint.

Fix this by changing the voltage settings to 1.15 V
(nominal) for both S1 and L3.

Note that 1.15 V is the only safely usable voltage
setting for VDD_MX and VDD_CX that can be used
without RPM support.

Also, fix a phandle for pm8026_s1_corner by renaming it to
pm8226_s1_corner to adequately reflect the platform's
name change.

Change-Id: I0d30fc1a37d12a355f9ea442482fbaf2d012cffe
Signed-off-by: David Keitel <dkeitel@codeaurora.org>