msm: smd: Add remote spinlock init check
For targets where other processors boot before the Apps processor, there
is a small risk that one of the processors will fail with the SMEM
remote spinlock locked.  In this case, the Apps will spin forever in SMD
init trying to acquire the spinlock.  Subsystem restart is not fully
functional until SMD has been initialized, so the Apps is not able to
recover.  It is important to note that this failure scenario should only
occur during development or due to a hardware failure.

Add a check to verify that the remote spinlock can be acquired during
SMD initialization.  If it cannot be acquired, then panic with a message
indicating which processor is not releasing the remote spinlock which
will help debug the root cause of the issue.

Change-Id: Idf4b453da5d84fd59f7787820ab0f2ab7bb6dba6
Signed-off-by: Eric Holmberg <eholmber@codeaurora.org>