msm: pil-q6v5-lpass: Assert LPASS_RESTART when shutting down LPASS
With LPASS clock control recently moved to the LPASS Q6, it's now
possible to do a full reset of the LPASS subsystem during
pil_lpass_shutdown(). When the reset is cleared again in pil_lpass_reset(),
the subsystem will have been completely reset to a clean state. This
increases the likelihood of the LPASS fully recovering after a subsystem
restart.

CRs-Fixed: 455368
Change-Id: I1da22c9aca00b1dffc1be3785a6beef3ed3359f7
Signed-off-by: Matt Wagantall <mattw@codeaurora.org>