iommu: msm: Refactor ASID assignment
Each context bank has an assigned ASID. Context banks sharing a page
table use the same ASID.

Instead of reading the ASID assigned to context banks store the
ASID in an array with a usage count. The index into the array is
the ASID value - 1. ASID assignments starts at 1. The number of
possible ASID values is equal to the number of context banks.

Change-Id: Ia332875505338b79a0b35679b92faf0a5aa2db84
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>