board: 8226: add SD card detect gpiomux configuration
On MSM8226 platforms, a MSM GPIO is used for SD card detect and it's
gpiomux configuration needs to be set to a specified value for the SD
card to be detected across all platforms. This patch adds the correct
gpiomux configuration for SD card detect GPIO to properly detect SD
card across all platforms.

Change-Id: I8bbc152b83087c52aeed3c0422390137fb56a9a4
Signed-off-by: Krishna Konda <kkonda@codeaurora.org>